pyproject.toml
[tool.poetry]
name = "helyos_agent_sdk"
version = "0.1.3"
description = ""
authors = ["Carlos Viol Barbosa <you@example.com>"]
readme = "README.md"
packages = [{include = "helyos_agent_sdk"}]
keywords = ["autonomous driving", "helyos", "sdk"]

homepage = "https://helyos.ivi.fraunhofer.de"
repository = "https://github.com/"
documentation = "https://helyos_agent_sdk.readthedocs.io/en/latest/"

classifiers = [
    "Environment :: Console",
    "Operating System :: OS Independent",
    "Topic :: Software Development :: Build Tools",
    "Topic :: Software Development :: Libraries :: Python Modules",
]


[tool.poetry.dependencies]
python = "^3.7"
pika = "^1.3.1"
pycryptodome = "^3.15.0"
dataclasses-json = "^0.5.7"


[tool.poetry.dev-dependencies]
pytest = "^5.2"
black = "^19.10b0"
pre-commit = {version = "^2.4.0", python = "^3.6.1"}
pylint = "^2.5.0"
tox = "^3.14.6"
sphinx = "^3.0.3"
sphinxcontrib-apidoc = "^0.3.0"
sphinx_rtd_theme = "^0.4.3"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
